refactor: 채팅방 참가 기능 분리
