+++ src/main/java/com/samsamhajo/deepground/chat/service/ChatRoomService.java
@@ -6,7 +6,6 @@ import com.samsamhajo.deepground.chat.entity.ChatRoomType;
 import com.samsamhajo.deepground.chat.repository.ChatRoomMemberRepository;
 import com.samsamhajo.deepground.chat.repository.ChatRoomRepository;
 import com.samsamhajo.deepground.member.entity.Member;
-import java.time.LocalDateTime;
 import java.util.Arrays;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -19,7 +18,6 @@ public class ChatRoomService {
 
     private final ChatRoomRepository chatRoomRepository;
     private final ChatRoomMemberRepository chatRoomMemberRepository;
-    private final ChatRedisService chatRedisService;
 
     @Transactional
     public ChatRoom createFriendChatRoom(Member member1, Member member2) {
@@ -46,18 +44,5 @@ public class ChatRoomService {
                 .map(member -> ChatRoomMember.of(member, chatRoom))
                 .toList();
         chatRoomMemberRepository.saveAll(chatRoomMembers);
-    }
-
-    @Transactional
-    public void joinChatRoom(Member member, ChatRoom chatRoom) {
-        ChatRoomMember chatRoomMember = ChatRoomMember.of(member, chatRoom);
-
-        // 채팅방에 메시지가 있다면 마지막으로 읽은 메시지 시간을 업데이트
-        LocalDateTime latestMessageTime = chatRedisService.getLatestMessageTime(chatRoom.getId());
-        if (latestMessageTime != null) {
-            chatRoomMember.updateLastReadMessageTime(latestMessageTime);
-        }
-
-        chatRoomMemberRepository.save(chatRoomMember);
     }
 }
